requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
requests==2.31.0
numpy==1.24.3
scikit-learn==1.3.0
python-dotenv==1.0.0
Flask-Cors==4.0.0
psycopg2-binary==2.9.7
redis==4.6.0
Werkzeug==2.3.7 